\name{disco-package}
\Rdversion{1.1}
\alias{disco-package}
\alias{disco}
\docType{package}
\title{
Comparing Distributions
}
\description{
This package contains a set of methods for 2-sample and K-sample distribution comparisons.
}
\details{
\tabular{ll}{
Package: \tab disco\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-02-07\cr
License: \tab GPL (>2)\cr
}
The package accompanies the book Comparing Distributions. All examples are given there.
}

\author{
Olivier Thas and Bert de Boeck

Maintainer: Joris Meys <Joris.Meys@UGent.be>
}

\references{
Comparing Distributions. Olivier Thas.
}

\keyword{ distributions, smoothtest }
\seealso{

}
\examples{

}
