% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{disco}
\alias{disco}
\alias{disco_palette_names}
\alias{disco_palette_data}
\title{Retrieve palette by name from palettes list stored in sysdata.rda}
\usage{
disco(palette, n = NULL, alpha = NULL, direction = 1)

disco_palette_names()

disco_palette_data(palette)
}
\arguments{
\item{palette}{The name of the palette from which to select colors}

\item{n}{Number of colors selected}

\item{alpha}{Degree of opacity in [0, 1]}

\item{direction}{Order of colors (1 = forward, -1 = reverse)}
}
\description{
Retrieve palette by name from palettes list stored in sysdata.rda
}
\keyword{internal}
