% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{disco_palette_show}
\alias{disco_palette_show}
\title{Show the different disco palettes}
\usage{
disco_palette_show(palettes = disco_palette_names(), n = NULL,
  alpha = NULL, direction = 1)
}
\arguments{
\item{palettes}{One or more palette names to show}

\item{n}{Number of colors selected}

\item{alpha}{Degree of opacity in [0, 1]}

\item{direction}{Order of colors (1 = forward, -1 = reverse)}
}
\description{
This is a simple function to show the selection of palettes
available in the `disco` package
}
\examples{

disco_palette_show()

}
