% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_disco}
\alias{scale_colour_disco}
\alias{scale_color_disco}
\alias{scale_fill_disco}
\title{A discrete scale for the colour aesthetic in ggplot2}
\usage{
scale_colour_disco(..., palette = "bright", alpha = NULL, direction = 1,
  aesthetics = "colour")

scale_color_disco(..., palette = "bright", alpha = NULL, direction = 1,
  aesthetics = "colour")

scale_fill_disco(..., palette = "bright", alpha = NULL, direction = 1,
  aesthetics = "fill")
}
\arguments{
\item{...}{Arguments to pass on to `ggplot2::discrete_scale()`}

\item{palette}{The name of the palette from which to select colors}

\item{alpha}{Degree of opacity in [0, 1]}

\item{direction}{Order of colors (1 = forward, -1 = reverse)}
}
\value{
A `ScaleDiscrete` object that can be added to a `ggplot` object
}
\description{
A discrete scale for the colour aesthetic in ggplot2

A discrete scale for the fill aesthetic in ggplot2
}
