% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{makeMatchedOrtholog}
\alias{makeMatchedOrtholog}
\title{Create an object of class matchedOrtholog}
\usage{
makeMatchedOrtholog(names, genes, lfc1, lfc2, pval1, pval2, row.names = NULL,
  extra = NULL)
}
\arguments{
\item{names}{character vector with names of the matched sets}

\item{genes}{names associated with the ortholog pairs}

\item{lfc1, lfc2}{log fold changes or effect sizes}

\item{pval1, pval2}{vectors of p values for the two data sets}

\item{row.names}{character vector to be used as row names for the data set}

\item{extra}{a data frame or a vector with the same length or number of rows as "genes"}
}
\description{
Create an object representing regulation in two sets of genes (e.g. from
two organisms).
}
\details{
The object contains a mapping between two sets of genes and the
associated effect sizes (i.e. log fold changes) and p-values used to
calculate disco score and other statistics.
}
\seealso{
matchedOrtholog-class
}
