% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{matchedOrtholog-class}
\alias{matchedOrtholog-class}
\alias{matchedOrtholog-class}
\alias{show,matchedOrtholog-method}
\alias{as}
\title{S4 class for disco}
\usage{
\S4method{show}{matchedOrtholog}(object)
}
\arguments{
\item{object}{An object of the matchedOrtholog class}
}
\description{
S4 class for disco
}
\details{
Class for holding matched orthologs and associated data. See
\code{makeMatchedOrholog} for details.
}
\seealso{
makeMatchedOrholog
}
