% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{pal11_index}
\alias{pal11_index}
\title{Retrieve the indices of colors in the 11-length palettes
based on the number of colors selected.}
\usage{
pal11_index(n)
}
\arguments{
\item{n}{Number of colors to select}
}
\description{
Retrieve the indices of colors in the 11-length palettes
based on the number of colors selected.
}
\keyword{internal}
