% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{pal9_index}
\alias{pal9_index}
\title{Retrieve the indices of colors in the 9-length palettes
based on the number of selected colors.}
\usage{
pal9_index(n)
}
\arguments{
\item{n}{Number of colors to select}
}
\description{
Retrieve the indices of colors in the 9-length palettes
based on the number of selected colors.
}
\keyword{internal}
