% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{rainbow_index}
\alias{rainbow_index}
\title{Retrieve the indices of colors in the rainbow palette
based on the number of colors selected.}
\usage{
rainbow_index(n)
}
\arguments{
\item{n}{Number of colors to select}
}
\description{
Retrieve the indices of colors in the rainbow palette
based on the number of colors selected.
}
\keyword{internal}
