% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signal.R
\name{Signal}
\alias{Signal}
\title{R6 abstract class defining a signal}
\description{
Defines an abstract interface for an R6 class that manages
  a multivariate time series data set. Full functional implementations
  of this interface must implement the functions described below.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time}}{A vector representing the times for the elements of the signal}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getWindow}{\code{Signal$getWindow()}}
\item \href{#method-getVariable}{\code{Signal$getVariable()}}
\item \href{#method-plotSummary}{\code{Signal$plotSummary()}}
\item \href{#method-plot}{\code{Signal$plot()}}
\item \href{#method-addVariable}{\code{Signal$addVariable()}}
\item \href{#method-writeCSV}{\code{Signal$writeCSV()}}
\item \href{#method-interpolate}{\code{Signal$interpolate()}}
\item \href{#method-clone}{\code{Signal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWindow"></a>}}
\if{latex}{\out{\hypertarget{method-getWindow}{}}}
\subsection{Method \code{getWindow()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Method for getting a subset of a signal. The time
  stamps included in the window should start on or after
  the minimum time and before or on the maximum time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$getWindow(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{minTime}}{Time defining the beginning of the window}

\item{\code{maxTime}}{Time defining the end of the window}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A subset of the signal as another Signal object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVariable"></a>}}
\if{latex}{\out{\hypertarget{method-getVariable}{}}}
\subsection{Method \code{getVariable()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Gets the vector of values for a given variable
  in the multivariate signal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$getVariable(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{variableName}}{The name of the variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSummary"></a>}}
\if{latex}{\out{\hypertarget{method-plotSummary}{}}}
\subsection{Method \code{plotSummary()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Plot a summary of the variables in the signal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$plotSummary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{x}}{X-axis variable}

\item{\code{mfrow}}{Plot layout (see \code{\link{par}})}

\item{\code{mar}}{Plot margins (see \code{\link{par}})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Plot a single variable in the signal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{variableName}}{The variable header to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addVariable"></a>}}
\if{latex}{\out{\hypertarget{method-addVariable}{}}}
\subsection{Method \code{addVariable()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Add a variable to a signal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$addVariable(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{variableHeader}}{The header for the property name for the new variable}

\item{\code{metadata}}{The metadata for the column}

\item{\code{values}}{The vector of values for the new column}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-writeCSV"></a>}}
\if{latex}{\out{\hypertarget{method-writeCSV}{}}}
\subsection{Method \code{writeCSV()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Write the signal in csv format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$writeCSV(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{path}}{Path to the CSV files that will be created}

\item{\code{name}}{Name base for the files}

\item{\code{timeVariableName}}{Header to use for the time column}

\item{\code{variables}}{Vector of variable names to include in output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-interpolate"></a>}}
\if{latex}{\out{\hypertarget{method-interpolate}{}}}
\subsection{Method \code{interpolate()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Creates a new signal by interpolating between points of
  this signal based on a provided vector of times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$interpolate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{time}}{Vector of times at which the signal should be interpolated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
