% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalDerivation.R
\name{SignalDerivation}
\alias{SignalDerivation}
\title{R6 class defining a signal derivation}
\description{
This defines an abstract interface to a signal derivation. A signal derivation
  is an analysis of a signal at a single location.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signal}}{The signal providing the basis for the derivation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SignalDerivation$new()}}
\item \href{#method-derive}{\code{SignalDerivation$derive()}}
\item \href{#method-clone}{\code{SignalDerivation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the class SignalDerivation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalDerivation$new(signal = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal}}{An optional argument allowing the derivation to be tied to a specific signal,
if desired.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-derive"></a>}}
\if{latex}{\out{\hypertarget{method-derive}{}}}
\subsection{Method \code{derive()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Performs a signal derivation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalDerivation$derive(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{signal}}{The signal on which the derivation is performed.
This parameter should only be required if the SignalDerivation object is
not already associated with a signal.}

\item{\code{prevResults}}{Object representing the results from a previous derivation (if applicable)}

\item{\code{path}}{The path where the results file is written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object representing the results of the derivation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalDerivation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
