% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableDataFile.R
\name{TableDataFile}
\alias{TableDataFile}
\title{A class for tabular data in a text data file (R6 Class)}
\description{
Provides utilities for managing tabular data stored in
  a text file using column delimiters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{instrument}}{Character string identifying the instrument used.}

\item{\code{filePath}}{character string representing the path to the file}

\item{\code{delimiter}}{Character string representing the delimiter for columns}

\item{\code{timeZone}}{character string representing the time zone (system dependent)}

\item{\code{timeFormat}}{character string representing the time format (strptime)}

\item{\code{timeVariableName}}{character string representing the time variable header in the table}

\item{\code{signal}}{signal containing data from the file}

\item{\code{numMetaRows}}{integer representing the number of rows of metadata at the top
of the file}

\item{\code{metaColumns}}{Data frame with context for each column}

\item{\code{metadata}}{Data frame with context for each column}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TableDataFile$new()}}
\item \href{#method-read}{\code{TableDataFile$read()}}
\item \href{#method-addMetadata}{\code{TableDataFile$addMetadata()}}
\item \href{#method-clone}{\code{TableDataFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableDataFile$new(
  instrument,
  filePath,
  timeVariableName,
  numMetaRows,
  metaColumns,
  delimiter,
  timeZone,
  timeFormat,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument}}{Character string identifying the instrument used}

\item{\code{filePath}}{Character string representing the path to the file}

\item{\code{timeVariableName}}{Character string representing the time variable header in the table.}

\item{\code{numMetaRows}}{Integer representing the number of rows of metadata at the top
of the file}

\item{\code{metaColumns}}{Data frame with context for each column.}

\item{\code{delimiter}}{Character string representing the delimiter for columns}

\item{\code{timeZone}}{Character string representing the time zone.}

\item{\code{timeFormat}}{character string representing the time format (strptime)}

\item{\code{metadata}}{Optional metadata object.
Defaults to NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Read the data file to populate the signal attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableDataFile$read(dataLayer = "root", stringsAsFactors = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stringsAsFactors}}{Optional logical value to turn strings as factors on or off.
Default value is FALSE (do not interpret strings as categorical factors).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The signal object created by reading the file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-addMetadata}{}}}
\subsection{Method \code{addMetadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableDataFile$addMetadata(metadata, instrument)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableDataFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
