% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferFunctionDerivation.R
\name{TransferFunctionDerivation}
\alias{TransferFunctionDerivation}
\title{R6 Class defining a transfer function derivation}
\description{
This defines an abstract interface to a transfer function derivation. 
  A transfer function derivation is an analysis of the changes made between
  an input and outpu signal.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signalIn}}{The input signal providing the basis for the derivation}

\item{\code{signalOut}}{The output signal providing the basis for the derivation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransferFunctionDerivation-new}{\code{TransferFunctionDerivation$new()}}
\item \href{#method-TransferFunctionDerivation-derive}{\code{TransferFunctionDerivation$derive()}}
\item \href{#method-TransferFunctionDerivation-clone}{\code{TransferFunctionDerivation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunctionDerivation-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunctionDerivation-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of class TransferFunctionDerivation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionDerivation$new(signalIn = NULL, signalOut = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signalIn}}{An optional argument allowing the derivation to be tied to a specific input signal,
if desired.}

\item{\code{signalOut}}{An optional argument allowing the derivation to be tied to a specific output signal,
if desired.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunctionDerivation-derive"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunctionDerivation-derive}{}}}
\subsection{Method \code{derive()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Performs a transfer function derivation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionDerivation$derive(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{signalIn}}{The input signal on which the derivation is performed.
This parameter is only required if the TransferFunctionDerivation object is
not already associated with an input signal 
(see \code{\link{TransferFunctionDerivation}})}

\item{\code{signalOut}}{The output signal on which the derivation is performed.
This parameter is only required if the TransferFunctionDerivation object is
not already associated with an output signal 
(see \code{\link{TransferFunctionDerivation}})}

\item{\code{prevResults}}{Object representing the results from a previous derivation (if applicable)}

\item{\code{path}}{The path where the results file is written}

\item{\code{index}}{Index of the derivation being performed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object representing the results of the derivation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunctionDerivation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunctionDerivation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionDerivation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
