% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferFunctionPlotter.R
\name{TransferFunctionPlotter}
\alias{TransferFunctionPlotter}
\title{R6 class defining a transfer function plotter}
\description{
Abstract class for providing visualizations of the 
  results of a tranfer function analysis.
  This class provides the definition of an interface to a
  TransferFunctionPlotter object and is not intended to be instantiated
  directly.
}
\section{Super class}{
\code{\link[disco:TransferFunctionSummarizer]{disco::TransferFunctionSummarizer}} -> \code{TransferFunctionPlotter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signalIn}}{The input signal to be plotted}

\item{\code{signalOut}}{The output signal to be plotted}

\item{\code{outputPath}}{The path containing the output}

\item{\code{fileName}}{The name of the pdf file to be generated}

\item{\code{mfrow}}{The structure of the multipanel plot}

\item{\code{mar}}{The margins for the plot}

\item{\code{device}}{The device to use for coding. If NULL, a device
will not be opened. A PDF device is currently the
only other option.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TransferFunctionPlotter$new()}}
\item \href{#method-open}{\code{TransferFunctionPlotter$open()}}
\item \href{#method-close}{\code{TransferFunctionPlotter$close()}}
\item \href{#method-clone}{\code{TransferFunctionPlotter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="disco" data-topic="TransferFunctionSummarizer" data-id="summarize">}\href{../../disco/html/TransferFunctionSummarizer.html#method-summarize}{\code{disco::TransferFunctionSummarizer$summarize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionPlotter$new(
  signalIn = NULL,
  signalOut = NULL,
  outputPath = NULL,
  fileName = "windowSummary.pdf",
  mfrow = c(3, 2),
  mar = c(4, 4, 2, 4) + 0.1,
  device = "pdf"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signalIn}}{Optional input signal associated with the plotter a priori.
Defaults to NULL.}

\item{\code{signalOut}}{Optional input signal associated with the plotter a priori.
Defaults to NULL.}

\item{\code{outputPath}}{Optional path to results from a dignal analysis.
Defaults to NULL.}

\item{\code{fileName}}{The file name for the pdf file generated
Defaults to "windowSummary.pdf"}

\item{\code{mfrow}}{The dimension of panels for the figure (see \code{\link{par}})
Defaults to c(3,2), or 3 rows and 2 columns of panels per figure.}

\item{\code{mar}}{The margins of each panel in the figure (see \code{\link{par}})
Defaults to c(4, 4, 2, 4) + 0.1}

\item{\code{device}}{The device to use for coding. If NULL, a device
will not be opened. A PDF device is currently the
only other option.
Defaults to "pdf".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\if{latex}{\out{\hypertarget{method-open}{}}}
\subsection{Method \code{open()}}{
Opens the PDF graphics device on the provided path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionPlotter$open(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{the path to which pdf files should be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
Closes the graphics device to write the PDF file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionPlotter$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
