% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferFunctionSummarizer.R
\name{TransferFunctionSummarizer}
\alias{TransferFunctionSummarizer}
\title{R6 class defining a transfer function summarizer}
\description{
Summarizes a transfer function and possibly its analysis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransferFunctionSummarizer-open}{\code{TransferFunctionSummarizer$open()}}
\item \href{#method-TransferFunctionSummarizer-summarize}{\code{TransferFunctionSummarizer$summarize()}}
\item \href{#method-TransferFunctionSummarizer-close}{\code{TransferFunctionSummarizer$close()}}
\item \href{#method-TransferFunctionSummarizer-clone}{\code{TransferFunctionSummarizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunctionSummarizer-open"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunctionSummarizer-open}{}}}
\subsection{Method \code{open()}}{
Performs an initial operations necessary to get the summarizer
  ready for multiple calls to method summarize.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionSummarizer$open(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{path}}{The path to the summarizer output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunctionSummarizer-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunctionSummarizer-summarize}{}}}
\subsection{Method \code{summarize()}}{
Perform the summary on an analysis of a provided 
  input and output signal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionSummarizer$summarize(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{signalIn}}{A multivariate input signal of type \code{\link{Signal}}}

\item{\code{signalOut}}{A multivariate output signal of type \code{\link{Signal}}}

\item{\code{outputPath}}{A path to results from a signal analysis.
Not necessary if summarizer does not need output.}

\item{\code{label}}{A label for the summary}

\item{\code{timeBounds}}{The temporal bounds on the summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return value not required.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunctionSummarizer-close"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunctionSummarizer-close}{}}}
\subsection{Method \code{close()}}{
Perform any operations necesary to close the summarizer
  and write any final output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionSummarizer$close(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return value not required.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransferFunctionSummarizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransferFunctionSummarizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransferFunctionSummarizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
