% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTable.R
\name{DataTable}
\alias{DataTable}
\title{An tabular data set with metadata}
\description{
An R6 implementation of a data table, including metadata
  for the data in the table
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{the S3 dataframe underlying the data structure}

\item{\code{meta}}{the S3 dataframe containing the general metadata}

\item{\code{metaColumns}}{the S3 dataframe containing the column metadata}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataTable$new()}}
\item \href{#method-constructFromCSV}{\code{DataTable$constructFromCSV()}}
\item \href{#method-copyMetaData}{\code{DataTable$copyMetaData()}}
\item \href{#method-addColumn}{\code{DataTable$addColumn()}}
\item \href{#method-clone}{\code{DataTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of a data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$new(dataframe, meta = NULL, metaColumns = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataframe}}{An R S3 dataframe object with the data.}

\item{\code{meta}}{Optional R S3 dataframe with general metadata.
Default value is NULL.}

\item{\code{metaColumns}}{Optional RS3 dataframe with column-specific metadata.
Default value is NULL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-constructFromCSV"></a>}}
\if{latex}{\out{\hypertarget{method-constructFromCSV}{}}}
\subsection{Method \code{constructFromCSV()}}{
Static method for creating a DataTable object 
  based on csv formatted information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$constructFromCSV(filePath, metaPath = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filePath}}{Path to the data file}

\item{\code{metaPath}}{Optional path to the metadata file.
Defaults to NULL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copyMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-copyMetaData}{}}}
\subsection{Method \code{copyMetaData()}}{
Copies the metadata from the provided data table 
  object to this data table object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$copyMetaData(dataTable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataTable}}{The data table object from which the metadata is copied}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addColumn"></a>}}
\if{latex}{\out{\hypertarget{method-addColumn}{}}}
\subsection{Method \code{addColumn()}}{
Adds a column of data and associated metata 
  as a variable to the data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$addColumn(header, metadata, values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{header}}{The header for the property name for the new column}

\item{\code{metadata}}{The metadata for the column}

\item{\code{values}}{The vector of values for the new column}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
