% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DielWindowedTransferFunction.R
\name{DielWindowedTransferFunction}
\alias{DielWindowedTransferFunction}
\title{Windowed transfer function analysis delimited by times of day}
\description{
Manages a windowed transfer function analysis based on an input
  and output signal, where the windows are 
  described by a shifting period of time starting at a given
  time of day, ending at a given time of day, and with a duration
  of a defined number of days apart.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signalIn}}{The input signal containing the data to be windowed}

\item{\code{signalOut}}{The output signal containing the data to be windowed}

\item{\code{path}}{The directory path to the windows}

\item{\code{dateStart}}{The start date for the windows}

\item{\code{dateEnd}}{The end date for the windows}

\item{\code{windows}}{Vector of dates representing the start of each window}

\item{\code{windowStart}}{The time of day each window starts}

\item{\code{windowEnd}}{The time of day each window ends}

\item{\code{windowDays}}{The number of days composing each window}

\item{\code{slideDays}}{The number of days to slide the window for generating each
window for analysis.}

\item{\code{inputPaths}}{The vector of paths to the input directories for each window}

\item{\code{inputFilePaths}}{The vector of paths to the input signal files for each window}

\item{\code{outputPaths}}{The vector of paths to the output directories for each window}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DielWindowedTransferFunction$new()}}
\item \href{#method-generateWindowInputOffset}{\code{DielWindowedTransferFunction$generateWindowInputOffset()}}
\item \href{#method-analyze}{\code{DielWindowedTransferFunction$analyze()}}
\item \href{#method-summarizeWindows}{\code{DielWindowedTransferFunction$summarizeWindows()}}
\item \href{#method-clone}{\code{DielWindowedTransferFunction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class DielWindowedTransferFunction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DielWindowedTransferFunction$new(
  signalIn,
  signalOut,
  path,
  dateStart,
  dateEnd,
  windowStart,
  windowEnd,
  windowDays,
  slideDays = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signalIn}}{The input signal object from which the windows are extracted}

\item{\code{signalOut}}{The output signal object from which the windows are extracted}

\item{\code{path}}{The path to the analysis input/output on the file system}

\item{\code{dateStart}}{The date to start the first window}

\item{\code{dateEnd}}{The date on which to end the last window}

\item{\code{windowStart}}{The time of day to start a given window}

\item{\code{windowEnd}}{The time of day to end a given window}

\item{\code{windowDays}}{The number of days for a given window}

\item{\code{slideDays}}{The number of days to slide the window for generating each
window for analysis.
Defaults to 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateWindowInputOffset"></a>}}
\if{latex}{\out{\hypertarget{method-generateWindowInputOffset}{}}}
\subsection{Method \code{generateWindowInputOffset()}}{
Parses the signal into windows based on the window definitions
  and a temporal offset between the output signal and the input signal.
  The input signal is interpolated to be a constant offset time from
  the output signal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DielWindowedTransferFunction$generateWindowInputOffset(
  offsetTime,
  pipelineDir = "dates",
  inputDir = "input"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{offsetTime}}{The offset time between the input signal and output signal in seconds}

\item{\code{pipelineDir}}{Optional directory structure from the base path where the folders corresponding
to analysis windows should be placed.
Default value is "dates".}

\item{\code{inputDir}}{Optional name of the directory within each window folder where the 
input data for that window should be placed.
Default value is "input".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-analyze"></a>}}
\if{latex}{\out{\hypertarget{method-analyze}{}}}
\subsection{Method \code{analyze()}}{
Runs the provided analyzer object for all windows
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DielWindowedTransferFunction$analyze(
  transferFunctionDerivation,
  pipelineDir = "dates",
  outputDir = "output"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transferFunctionDerivation}}{A TransferFunctionDerivation object that will perform the analysis
for each window}

\item{\code{pipelineDir}}{Optional directory structure from the base path where the folders corresponding
to analysis windows should be placed.
Default value is "dates".}

\item{\code{outputDir}}{Optional name of the directory within each window folder where the 
output data for that window should be placed.
Default value is "output".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarizeWindows"></a>}}
\if{latex}{\out{\hypertarget{method-summarizeWindows}{}}}
\subsection{Method \code{summarizeWindows()}}{
Uses the provided TransferFunctionSummarizer to generate a summary
  of the analysis of each window.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DielWindowedTransferFunction$summarizeWindows(
  transferFunctionSummarizer,
  summaryDir = "summary",
  useResults = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transferFunctionSummarizer}}{A transfer function summarizer that will generate the summary output for each
window}

\item{\code{summaryDir}}{Optional path from the base path where the summary output will be written.
Default value is "summary".}

\item{\code{useResults}}{Optional switch to turn of summarization of results.
Default value is TRUE (results will be sumarized).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DielWindowedTransferFunction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
