% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-tables.R
\name{gt_uni_separation}
\alias{gt_uni_separation}
\title{Create a gt table for a univariate result (no overlap column)}
\usage{
gt_uni_separation(
  res,
  title = "Univariate Separation",
  subtitle = NULL,
  digits = 3,
  show_rows_used = FALSE
)
}
\arguments{
\item{res}{Result of uni_separation()}

\item{title, subtitle}{Title/subtitle strings}

\item{digits}{Numeric formatting decimals}

\item{show_rows_used}{If TRUE, show a preview of row indices}
}
\description{
Create a gt table for a univariate result (no overlap column)
}
