% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{Metadata}
\alias{Metadata}
\title{Metadata for a data product (R6 Class)}
\description{
Provides utilities for managing metadata for a data product
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Metadata-new}{\code{Metadata$new()}}
\item \href{#method-Metadata-createLayerWithMetadata}{\code{Metadata$createLayerWithMetadata()}}
\item \href{#method-Metadata-writeXML}{\code{Metadata$writeXML()}}
\item \href{#method-Metadata-readXML}{\code{Metadata$readXML()}}
\item \href{#method-Metadata-getLayer}{\code{Metadata$getLayer()}}
\item \href{#method-Metadata-addMetadataElement}{\code{Metadata$addMetadataElement()}}
\item \href{#method-Metadata-addPerson}{\code{Metadata$addPerson()}}
\item \href{#method-Metadata-addVariable}{\code{Metadata$addVariable()}}
\item \href{#method-Metadata-addEntity}{\code{Metadata$addEntity()}}
\item \href{#method-Metadata-addEntityAtSite}{\code{Metadata$addEntityAtSite()}}
\item \href{#method-Metadata-addFile}{\code{Metadata$addFile()}}
\item \href{#method-Metadata-addTimeWindow}{\code{Metadata$addTimeWindow()}}
\item \href{#method-Metadata-createSiteElement}{\code{Metadata$createSiteElement()}}
\item \href{#method-Metadata-addSite}{\code{Metadata$addSite()}}
\item \href{#method-Metadata-createPointElement}{\code{Metadata$createPointElement()}}
\item \href{#method-Metadata-addPoint}{\code{Metadata$addPoint()}}
\item \href{#method-Metadata-addInstrument}{\code{Metadata$addInstrument()}}
\item \href{#method-Metadata-addQAQCCode}{\code{Metadata$addQAQCCode()}}
\item \href{#method-Metadata-addProjectEffort}{\code{Metadata$addProjectEffort()}}
\item \href{#method-Metadata-clone}{\code{Metadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$new(path, baseLayerName = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-createLayerWithMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-createLayerWithMetadata}{}}}
\subsection{Method \code{createLayerWithMetadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$createLayerWithMetadata(layer = NULL, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-writeXML"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-writeXML}{}}}
\subsection{Method \code{writeXML()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$writeXML()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-readXML"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-readXML}{}}}
\subsection{Method \code{readXML()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$readXML()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-getLayer"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-getLayer}{}}}
\subsection{Method \code{getLayer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$getLayer(layerName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addMetadataElement"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addMetadataElement}{}}}
\subsection{Method \code{addMetadataElement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addMetadataElement(element, category, context, description, layer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addPerson"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addPerson}{}}}
\subsection{Method \code{addPerson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addPerson(
  context,
  name,
  org,
  prefContact,
  prefContactType = "email",
  category = "who",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addVariable"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addVariable}{}}}
\subsection{Method \code{addVariable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addVariable(
  context,
  header,
  name,
  units,
  dimensions,
  stat,
  instrument,
  category = "what",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addEntity"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addEntity}{}}}
\subsection{Method \code{addEntity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addEntity(
  context,
  entityPath,
  category = "what",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addEntityAtSite"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addEntityAtSite}{}}}
\subsection{Method \code{addEntityAtSite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addEntityAtSite(
  entityContext,
  entityPath,
  siteContext,
  sitePath,
  addPoint = FALSE,
  pointContext = NULL,
  source = NULL,
  latitude = NULL,
  latitudeUnits = NULL,
  longitude = NULL,
  longitudeUnits = NULL,
  elevation = "unknown",
  elevationUnits = "unknown",
  entityCategory = "what",
  siteCategory = "where",
  pointCategory = "where",
  description = NULL,
  siteDescription = NULL,
  pointDescription = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addFile"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addFile}{}}}
\subsection{Method \code{addFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addFile(
  context,
  fileName,
  fileType,
  contentType,
  category = "what",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addTimeWindow"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addTimeWindow}{}}}
\subsection{Method \code{addTimeWindow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addTimeWindow(
  context,
  startTime,
  endTime,
  timeZone,
  category = "when",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-createSiteElement"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-createSiteElement}{}}}
\subsection{Method \code{createSiteElement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$createSiteElement(sitePath)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addSite"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addSite}{}}}
\subsection{Method \code{addSite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addSite(
  context,
  sitePath,
  category = "where",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-createPointElement"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-createPointElement}{}}}
\subsection{Method \code{createPointElement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$createPointElement(
  source,
  latitude,
  latitudeUnits,
  longitude,
  longitudeUnits,
  elevation = "unknown",
  elevationUnits = "unknown"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addPoint"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addPoint}{}}}
\subsection{Method \code{addPoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addPoint(
  context,
  source,
  latitude,
  latitudeUnits,
  longitude,
  longitudeUnits,
  elevation = "unknown",
  elevationUnits = "unknown",
  category = "where",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addInstrument"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addInstrument}{}}}
\subsection{Method \code{addInstrument()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addInstrument(
  context,
  name,
  serialNumber,
  category = "how",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addQAQCCode"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addQAQCCode}{}}}
\subsection{Method \code{addQAQCCode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addQAQCCode(
  context,
  header,
  code,
  category = "how",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-addProjectEffort"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-addProjectEffort}{}}}
\subsection{Method \code{addProjectEffort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$addProjectEffort(
  context,
  path,
  category = "why",
  description = NULL,
  layerName = NULL,
  layer = self$getLayer(layerName = layerName)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
