% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalPlotter.R
\name{SignalPlotter}
\alias{SignalPlotter}
\title{R6 class defining a signal plotter}
\description{
Abstract class for providing visualizations of a signal analysis.
}
\section{Super class}{
\code{\link[disco:SignalSummarizer]{disco::SignalSummarizer}} -> \code{SignalPlotter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signal}}{The signal to be plotted}

\item{\code{outputPath}}{The path containing the output}

\item{\code{fileName}}{The name of the pdf file to be generated}

\item{\code{mfrow}}{The structure of the multipanel plot}

\item{\code{mar}}{The margins for the plot}

\item{\code{device}}{The device to use for coding. If NULL, a device
will not be opened. A PDF device is currently the
only other option.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SignalPlotter-new}{\code{SignalPlotter$new()}}
\item \href{#method-SignalPlotter-open}{\code{SignalPlotter$open()}}
\item \href{#method-SignalPlotter-close}{\code{SignalPlotter$close()}}
\item \href{#method-SignalPlotter-clone}{\code{SignalPlotter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="disco" data-topic="SignalSummarizer" data-id="summarize"><a href='../../disco/html/SignalSummarizer.html#method-SignalSummarizer-summarize'><code>disco::SignalSummarizer$summarize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignalPlotter-new"></a>}}
\if{latex}{\out{\hypertarget{method-SignalPlotter-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class SignalPlotter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalPlotter$new(
  signal = NULL,
  outputPath = NULL,
  fileName = "windowSummary.pdf",
  mfrow = c(3, 2),
  mar = c(4, 4, 2, 4) + 0.1,
  device = "pdf"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal}}{Optional signal associated with the plotter a priori.
Defaults to NULL.}

\item{\code{outputPath}}{Optional path to results from a signal analysis.
Defaults to NULL.}

\item{\code{fileName}}{The file name for the pdf file generated
Defaults to "windowSummary.pdf"}

\item{\code{mfrow}}{The dimension of panels for the figure (see \code{\link{par}})
Defaults to c(3,2), or 3 rows and 2 columns of panels per figure.}

\item{\code{mar}}{The margins of each panel in the figure (see \code{\link{par}})
Defaults to c(4, 4, 2, 4) + 0.1}

\item{\code{device}}{The device to use for coding. If NULL, a device
will not be opened. A PDF device is currently the
only other option.
Defaults to "pdf".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignalPlotter-open"></a>}}
\if{latex}{\out{\hypertarget{method-SignalPlotter-open}{}}}
\subsection{Method \code{open()}}{
Opens the PDF graphics device on the provided path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalPlotter$open(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{the path to which pdf files should be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignalPlotter-close"></a>}}
\if{latex}{\out{\hypertarget{method-SignalPlotter-close}{}}}
\subsection{Method \code{close()}}{
Closes the graphics device to write the PDF file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalPlotter$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignalPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SignalPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
