% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalSummarizer.R
\name{SignalSummarizer}
\alias{SignalSummarizer}
\title{R6 class defining a signal summarizer}
\description{
Summarizes a signal and possibly its analysis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-open}{\code{SignalSummarizer$open()}}
\item \href{#method-summarize}{\code{SignalSummarizer$summarize()}}
\item \href{#method-close}{\code{SignalSummarizer$close()}}
\item \href{#method-clone}{\code{SignalSummarizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\if{latex}{\out{\hypertarget{method-open}{}}}
\subsection{Method \code{open()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Performs an initial operations necessary to get the summarizer
  ready for multiple calls to method summarize.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalSummarizer$open(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{path}}{The path to the summarizer output}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return value not required.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Perform the summary on a provided signal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalSummarizer$summarize(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{signal}}{The signal object to be analyzed}

\item{\code{outputPath}}{A path to results from a signal analysis.
Not necessary if summarizer does not need output.}

\item{\code{label}}{A label for the summary}

\item{\code{timeBounds}}{The temporal bounds on the summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return value not required.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
This is an abstract declaration and the method
  must be implemented by extending classes.
  
  Perform any operations necesary to close the summarizer
  and write any final output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalSummarizer$close(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return value not required.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignalSummarizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
